+++ project.yaml
@@ -35,3 +35,10 @@ actions:
     outputs:
       highly_sensitive:
         dataset: output/dataset_patients_debug.csv.gz
+
+  # action pending discussion - combine monthly tables
+  combine_monthly_patient_gz:
+    run: python preprocess_combine_gz.py --output output/dataset_patients_combined.csv
+    outputs:
+      highly_sensitive:
+        dataset: output/dataset_patients_combined.csv
